Refactor: 이메일 검사 페이지 컴포넌트 분리
