+++ src/components/checkEmail/styles.ts
